refactor: 전역 스토어 변경에 따른 기능 재배분
