+++ react-query/profile/useGetMyProfile.ts
@@ -5,8 +5,8 @@ import {
 } from '@tanstack/react-query';
 import useErrorStore from 'store/useErrorStore';
 import { MyInfo } from 'types/MyInfo';
-import { getMyProfile } from '../api/profile/Api';
-import QueryKey from './key';
+import { getProfile } from '../../api/profile/Api';
+import QueryKey from '../react-key';
 
 export const useGetMyProfile = (): UseMutateAsyncFunction<
   MyInfo,
@@ -17,7 +17,7 @@ export const useGetMyProfile = (): UseMutateAsyncFunction<
   const queryClient = useQueryClient();
   const { setError } = useErrorStore();
   const { mutateAsync } = useMutation({
-    mutationFn: (id: number) => getMyProfile(id),
+    mutationFn: (id: number) => getProfile(id),
     onError: () => {
       setError(true);
     },
