+++ react-query/profile/useUpdateProfile.ts
@@ -4,9 +4,9 @@ import {
   useQueryClient,
 } from '@tanstack/react-query';
 import { updateUserProfile } from 'api/profile/Api';
-import useErrorStore from 'store/useErrorStore';
-import QueryKey from './key';
+import QueryKey from '../react-key';
 import useDisabledStore from 'store/myPageFormStore';
+import sliderStore from 'store/sliderStore';
 
 type PatchUserInfo = {
   nickname: string;
@@ -21,7 +21,7 @@ export const useUpdateProfile = (): UseMutateFunction<
   unknown
 > => {
   const queryClient = useQueryClient();
-  const { setErrorUpdateChecker, setErrorSlider } = useErrorStore();
+  const { setUpdateCheckerSlider, setFormSlider } = sliderStore();
 
   const { setDisabledAll } = useDisabledStore();
   const { mutate } = useMutation({
@@ -29,12 +29,12 @@ export const useUpdateProfile = (): UseMutateFunction<
       updateUserProfile(nickname, 'avatars/' + avatarUrl, bio),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [QueryKey.getLoginedUser] });
-      setErrorUpdateChecker(true);
-      setErrorSlider(false); //업데이트하고 업데이트 여부 묻는 슬라이더 꺼준다
+      setUpdateCheckerSlider(true); //업데이트 하고 슬라이더 애니메이션 켜준다
+      setFormSlider(false); //업데이트하고 업데이트 여부 묻는 슬라이더 꺼준다
       setDisabledAll(); //업데이트하고 input disabled로 만든다
       setTimeout(() => {
-        //업데이트하고 슬라이더 애니메이션을 꺼준다
-        setErrorUpdateChecker(false);
+        //일정 시간 후 슬라이더 애니메이션을 꺼준다
+        setUpdateCheckerSlider(false);
       }, 2000);
     },
     onError: (error) => {
